feat: Update date filters in EmbarquesController.php
This commit updates the date filters in EmbarquesController.php to improve the accuracy of the query results. The code now includes the departure date in the WHERE clause to ensure that only relevant data is retrieved. This change enhances the functionality of the EmbarquesController and provides more precise filtering options for the users.