feat: Set default value for stock_ID in Lancamento.php
This commit sets a default value of 1 for the $stock_ID parameter in the Lancamento.php file. Previously, the code checked if $stock_ID was null and assigned it a value of 1 if it was. By setting the default value directly in the parameter declaration, the code becomes more concise and easier to understand. This change improves the readability and maintainability of the code.